+++ tests/test_pdp/test_pdp-with_redis.py
@@ -1,5 +1,5 @@
 """
-    PDP tests with MongoDb storage
+    PDP tests with Redis storage
 """
 
 import pytest
@@ -8,14 +8,13 @@ from py_abac.pdp import PDP, EvaluationAlgorithm
 from py_abac.policy import Policy
 from py_abac.provider.base import AttributeProvider
 from py_abac.request import AccessRequest
-from py_abac.storage.mongo import MongoStorage
-from ..test_storage.test_mongo import create_client
+from py_abac.storage.redis import RedisStorage
+from ..test_storage.test_redis import create_client
 
 # Pytest mark for module
-pytestmark = [pytest.mark.mongo, pytest.mark.integration]
+pytestmark = [pytest.mark.redis, pytest.mark.integration]
 
-DB_NAME = 'db_test'
-COLLECTION = 'policies_test'
+HASH_KEY = 'py_abac_policies_test'
 SUBJECT_IDS = {"Max": "user:1", "Nina": "user:2", "Ben": "user:3", "Henry": "user:4"}
 POLICIES = [
     {
@@ -160,11 +159,11 @@ class EmailsAttributeProvider(AttributeProvider):
 @pytest.fixture
 def st():
     client = create_client()
-    storage = MongoStorage(client, DB_NAME, collection=COLLECTION)
+    storage = RedisStorage(client, hash_key=HASH_KEY)
     for policy_json in POLICIES:
         storage.add(Policy.from_json(policy_json))
     yield storage
-    client[DB_NAME][COLLECTION].drop()
+    client.flushdb()
     client.close()